rename ants_system_v2 -> ants_system
